Create the base pages and route
